v2functions/requirements.txt
# Minimum packages for azure functions
azure-functions
azure-functions-worker
grpcio~=1.20.1
grpcio-tools~=1.20.1
protobuf==3.6.1
six==1.11.0
# Additional packages
requests==2.20.1
feedparser==5.2.1
Pillow==6.0.0
# Required for Azure Media Services
adal
msrest
azure-mgmt-media
azure-storage-blob
